requirements/base.txt
Django==4.0.0
djangorestframework==3.14
psycopg2==2.9.6
phonenumbers==8.13.20
django-phonenumber-field==7.1.0
